+++ api/AltV.Net.Client/Elements/Entities/ColShape.cs
@@ -28,7 +28,7 @@ namespace AltV.Net.Client.Elements.Entities
             ColShapeNativePointer = nativePointer;
         }
 
-        public bool IsEntityIdIn(ushort id)
+        public bool IsEntityIdIn(uint id)
         {
             CheckIfEntityExists();
